removed RoomAvailabilityBox component declared in BuildingDrawer
