+++ frontend/views/RoomAvailabilityBox.tsx
@@ -1,23 +1,50 @@
 import { RoomStatus } from "@common/types";
 import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import Box from "@mui/material/Box";
-import Typography from "@mui/material/Typography";
+import { Typography, TypographyProps } from "@mui/material";
+import Box, { BoxProps } from "@mui/material/Box";
+import { styled } from "@mui/material/styles";
 import Link from "next/link";
 import React from "react";
 
 import useRoom from "../hooks/useRoom";
 
-interface RoomAvailabilityBoxProps {
+const IndiviRoomBox = styled(Box)<BoxProps>(({ theme }) => ({
+  display: "flex",
+  flexDirection: "row",
+  justifyContent: "space-between",
+  alignItems: "center",
+  borderRadius: 10,
+  height: 70,
+  fontSize: 20,
+  fontWeight: 500,
+  backgroundColor: "#FFFFFF",
+  color: "black",
+  padding: theme.spacing(2, 2, 2, 3),
+  margin: theme.spacing(1.5, 1),
+  "&:hover": {
+    border: "1px solid",
+    borderColor: theme.palette.primary.main,
+    cursor: "pointer",
+  },
+}));
+
+const RoomBoxHeading = styled(Typography)<TypographyProps>(({ theme }) => ({
+  fontSize: 16,
+  fontWeight: 500,
+  whiteSpace: "nowrap",
+}));
+
+const RoomBoxSubheading = styled(Typography)<TypographyProps>(({ theme }) => ({
+  fontSize: 12,
+  fontWeight: 400,
+  paddingTop: 1,
+}));
+
+const RoomAvailabilityBox: React.FC<{
   roomNumber: string;
   roomStatus: RoomStatus;
   buildingId: string;
-}
-
-const RoomAvailabilityBox: React.FC<RoomAvailabilityBoxProps> = ({
-  roomNumber,
-  roomStatus,
-  buildingId,
-}) => {
+}> = ({ roomNumber, roomStatus, buildingId }) => {
   const date = new Date(roomStatus.endtime);
   const hoursMinutes = date.toLocaleTimeString("en-AU", {
     hour: "2-digit",
@@ -38,77 +65,55 @@ const RoomAvailabilityBox: React.FC<RoomAvailabilityBoxProps> = ({
   };
 
   const untilMessage = {
-    free: hoursMinutes === "Invalid Date" ? "" : `until ${hoursMinutes}`,
-    busy: hoursMinutes === "Invalid Date" ? "" : `until ${hoursMinutes}`,
-    soon: `at ${hoursMinutes}`,
+    free: hoursMinutes == "Invalid Date" ? "" : "until " + hoursMinutes,
+    busy: hoursMinutes == "Invalid Date" ? "" : "until " + hoursMinutes,
+    soon: "at " + hoursMinutes,
   };
 
   const { room } = useRoom(`${buildingId}-${roomNumber}`);
 
   return (
     <Link href={`/room/${buildingId}-${roomNumber}`}>
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "row",
-          justifyContent: "space-between",
-          alignItems: "center",
-          borderRadius: 10,
-          height: 70,
-          fontSize: 20,
-          fontWeight: 500,
-          backgroundColor: "#FFFFFF",
-          color: "black",
-          padding: "16px 24px",
-          margin: "12px 8px",
-          cursor: "pointer",
-          transition: "border-color 0.3s ease",
-          "&:hover": {
-            borderColor: "#1976d2",
-          },
-        }}
-      >
-        <Box>
-          <Typography sx={{ fontSize: 16, fontWeight: 500 }}>
-            {roomNumber}
-          </Typography>
-          <Typography sx={{ fontSize: 12, fontWeight: 400 }}>
-            {!room ? "" : room.name}
-          </Typography>
+      <IndiviRoomBox>
+        <Box
+          sx={{
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "flex-start",
+            paddingRight: 1,
+          }}
+        >
+          <RoomBoxHeading>{roomNumber}</RoomBoxHeading>
+          <RoomBoxSubheading>{!room ? "" : room.name}</RoomBoxSubheading>
         </Box>
         <Box
-          sx={{ display: "flex", flexDirection: "row", alignItems: "center" }}
+          sx={{
+            display: "flex",
+            flexDirection: "row",
+            justifyContent: "space-around",
+            alignItems: "center",
+          }}
         >
           <Box
             sx={{
               display: "flex",
               flexDirection: "column",
               alignItems: "flex-end",
-              marginRight: 1,
+              paddingRight: 1,
             }}
           >
-            <Typography
-              sx={{
-                color: roomStatusColor[roomStatus.status],
-                fontSize: 16,
-                fontWeight: 500,
-              }}
-            >
+            <RoomBoxHeading sx={{ color: roomStatusColor[roomStatus.status] }}>
               {roomStatusMessage[roomStatus.status]}
-            </Typography>
-            <Typography
-              sx={{
-                color: roomStatusColor[roomStatus.status],
-                fontSize: 12,
-                fontWeight: 400,
-              }}
+            </RoomBoxHeading>
+            <RoomBoxSubheading
+              sx={{ color: roomStatusColor[roomStatus.status] }}
             >
               {untilMessage[roomStatus.status]}
-            </Typography>
+            </RoomBoxSubheading>
           </Box>
           <ChevronRightIcon style={{ color: "grey" }} />
         </Box>
-      </Box>
+      </IndiviRoomBox>
     </Link>
   );
 };
